src/test/java/com/example/alumni/feature/share_comment/ShareCommentServiceImplTest.java
package com.example.alumni.feature.share_comment;

import com.example.alumni.feature.share.ShareRepository;
import com.example.alumni.feature.share_comment.*;
import com.example.alumni.feature.share_comment.dto.*;
import com.example.alumni.domain.*;
import com.example.alumni.feature.user.UserRepository;
import com.example.alumni.mapper.*;
import com.example.alumni.utils.*;
import org.junit.jupiter.api.*;
import org.mockito.*;
import org.springframework.data.domain.*;

import java.util.*;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.*;

public class ShareCommentServiceImplTest {

    @InjectMocks
    private ShareCommentServiceImpl shareCommentService;

    @Mock
    private ShareCommentRepository shareCommentRepository;

    @Mock
    private ShareRepository shareRepository;

    @Mock
    private UserRepository userRepository;

    @Mock
    private ShareCommentMapper shareCommentMapper;

    @BeforeEach
    public void setup() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    public void testCreateShareComment() {
        // Prepare data
        // TODO: Replace with actual instances
        ShareCommentRequest shareCommentRequest = new ShareCommentRequest("arg1", "arg2", "arg3");
        ShareComment shareComment = new ShareComment();
        User user = new User();
        Share share = new Share();
        List<ShareCommentResponse> list = new ArrayList<>(); // replace with actual list
        ShareCommentResponse shareCommentResponse = new ShareCommentResponse("arg1", "arg2", "arg3", list, "arg5", true);

        // Define the behavior of mocked objects
        when(shareCommentMapper.fromRequestToResponse(any())).thenReturn(shareComment);
        when(userRepository.findById(any())).thenReturn(Optional.of(user));
        when(shareRepository.findById(any())).thenReturn(Optional.of(share));
        when(shareCommentRepository.save(any())).thenReturn(shareComment);
        when(shareCommentMapper.toShareCommentResponse(any())).thenReturn(shareCommentResponse);

        // Call the method to be tested
        ShareCommentResponse result = shareCommentService.createShareComment(shareCommentRequest);

        // Verify the result
        assertEquals(shareCommentResponse, result);
    }

    @Test
    public void testUpdateShareComment() {
        // Prepare data
        // TODO: Replace with actual instances
        ShareCommentRequest shareCommentRequest = new ShareCommentRequest("arg1", "arg2", "arg3");
        ShareComment shareComment = new ShareComment();
        User user = new User();
        Share share = new Share();
        List<ShareCommentResponse> list = new ArrayList<>(); // replace with actual list
        ShareCommentResponse shareCommentResponse = new ShareCommentResponse("arg1", "arg2", "arg3", list, "arg5", true);

        // Define the behavior of mocked objects
        when(shareCommentRepository.findById(any())).thenReturn(Optional.of(shareComment));
        when(userRepository.findById(any())).thenReturn(Optional.of(user));
        when(shareRepository.findById(any())).thenReturn(Optional.of(share));
        when(shareCommentRepository.save(any())).thenReturn(shareComment);
        when(shareCommentMapper.toShareCommentResponse(any())).thenReturn(shareCommentResponse);

        // Call the method to be tested
        ShareCommentResponse result = shareCommentService.updateShareComment("shareCommentId", shareCommentRequest);

        // Verify the result
        assertEquals(shareCommentResponse, result);
    }

    @Test
    public void testDeleteShareComment() {
        // Define the behavior of mocked objects
        doNothing().when(shareCommentRepository).deleteById(any());

        // Call the method to be tested
        shareCommentService.deleteShareComment("shareCommentId");

        // Verify the behavior
        verify(shareCommentRepository, times(1)).deleteById(any());
    }


}
